+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+- focus on reveal
+
 ## [0.3.2] - 2021-05-18
 - yes/no are booleans (https://github.com/kubernetes/kubernetes/issues/34146)